Create views of sequences on the PlanetScale side before proxying there
The original implementation missed this step, thus requiring the final
application-level switch from the proxy to the real PlanetScale database
to happen very quickly after the proxy switch in order to be correct.

Now the FDW will read sequences from views on the new server after
`mv-proxy.sh` switches traffic to it.